Implement a choice type for Pystachio
We're looking to use Pystachio for configuring services at Dropbox. One problem that we came across
is that some of our configurations depend on attributes of the specific machine being deployed to. So we'd like to be able to have some config fields be either integers or an expression that can be evaluated on the host.

This change adds a generic choice type, which typechecks if its value is the value of any of a
list of choices.