Improve performance of passing a non-Mapping object to a Structural
Problem
-------

Serializing large configurations use a lot of CPU time. Profiling
revealed that about 50% of the CPU time was due to calling __repr__,
which seemed suspicious. This is because we are using repr() to
instantiate a ValueError that is immediately caught and discarded.

Solution
--------

Create a subclass of ValueError that lazily renders the message,
avoiding the call to repr in the hot path (ChoiceContainer._unwrap).